Add pallet-lending in runtime

+++ frame/oracle/src/mock.rs
@@ -123,7 +123,7 @@ impl pallet_oracle::Config for Test {
 	type Event = Event;
 	type AuthorityId = crypto::TestAuthId;
 	type Currency = Balances;
-	type AssetId = u64;
+	type AssetId = u128;
 	type PriceValue = u128;
 	type StakeLock = StakeLock;
 	type StalePrice = StalePrice;
